# src/api/rest-api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateProjectDto": {
        "properties": {
          "branchName": {
            "default": "master",
            "type": "string"
          },
          "projectName": {
            "type": "string"
          }
        },
        "required": [
          "projectName"
        ],
        "type": "object"
      },
      "CreateRevisionDto": {
        "properties": {
          "comment": {
            "default": "",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ErrorModel": {
        "properties": {
          "error": {
            "description": "Error type",
            "example": "Bad Request",
            "type": "string"
          },
          "message": {
            "description": "Detailed error message",
            "example": "Validation failed",
            "type": "string"
          },
          "statusCode": {
            "description": "HTTP status code",
            "example": 400,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
...
      "ProjectModel": {
        "properties": {
          "createdAt": {
            "description": "Timestamp when the project was created",
            "example": "2024-01-15T10:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Unique project identifier",
            "example": "V1StGXR8_Z5jdHi6B-myT",
            "type": "string"
          },
          "name": {
            "description": "URL-friendly project name (lowercase letters, numbers, hyphens)",
            "example": "my-blog",
            "type": "string"
          },
          "organizationId": {
            "description": "Organization that owns this project",
            "example": "acme-corp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organizationId",
          "createdAt",
          "name"
        ],
        "type": "object"
      },
...
      "ProjectsConnection": {
        "properties": {
          "edges": {
            "description": "List of items with cursors",
            "items": {
              "$ref": "#/components/schemas/ProjectModelEdgeType"
            },
            "type": "array"
          },
          "pageInfo": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PageInfo"
              }
            ],
            "description": "Pagination metadata"
          },
          "totalCount": {
            "description": "Total number of items available",
            "type": "number"
          }
        },
        "required": [
          "edges",
          "totalCount",
          "pageInfo"
        ],
        "type": "object"
      },
...
      "SuccessModelDto": {
        "properties": {
          "success": {
            "example": true,
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "UpdateProjectDto": {
        "properties": {
          "isPublic": {
            "type": "boolean"
          }
        },
        "required": [
          "isPublic"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "access-token": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "Revisium API",
    "version": "2.6.0-alpha.3"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/organization/{organizationId}/projects": {
      "get": {
        "operationId": "projects",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "first",
            "required": true,
            "schema": {
              "default": 100,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsConnection"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Organization not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List projects in organization",
        "tags": [
          "Organization"
        ]
      },
      "post": {
        "operationId": "createProject",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Clone from existing revision",
            "in": "query",
            "name": "fromRevisionId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Organization not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new project",
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/organization/{organizationId}/projects/{projectName}": {
      "delete": {
        "operationId": "deleteProject",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessModelDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a project",
        "tags": [
          "Project"
        ]
      },
      "get": {
        "operationId": "project",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get project by name",
        "tags": [
          "Project"
        ]
      },
      "put": {
        "operationId": "updateProject",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessModelDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update project settings",
        "tags": [
          "Project"
        ]
      }
    },
    "/api/organization/{organizationId}/projects/{projectName}/branches": {
      "get": {
        "operationId": "branches",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "first",
            "required": true,
            "schema": {
              "default": 100,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchesConnection"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List all branches in the project",
        "tags": [
          "Project"
        ]
      }
    },
    "/api/organization/{organizationId}/projects/{projectName}/branches/{branchName}": {
      "get": {
        "operationId": "branch",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Branch name",
            "in": "path",
            "name": "branchName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Branch not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get branch by name",
        "tags": [
          "Branch"
        ]
      }
    },
    "/api/organization/{organizationId}/projects/{projectName}/branches/{branchName}/create-revision": {
      "post": {
        "operationId": "createRevision",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Branch name",
            "in": "path",
            "name": "branchName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRevisionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Branch not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Commit changes and create a new revision",
        "tags": [
          "Branch"
        ]
      }
    },
    "/api/organization/{organizationId}/projects/{projectName}/branches/{branchName}/draft-revision": {
      "get": {
        "operationId": "draftRevision",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Branch name",
            "in": "path",
            "name": "branchName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Branch not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get the draft (working) revision for modifications",
        "tags": [
          "Branch"
        ]
      }
    },
...
  },
  "servers": [],
...
}